move native token type back
